+++ src/entry.ts
@@ -38,7 +38,12 @@ export function toEntryList(action: Pick<Action, 'fields'>): EntryList {
         appendCheckbox(entryList, field);
         break;
       default:
-        appendEntry(entryList, field.name, String(field.value ?? ''));
+        appendEntry(
+          entryList,
+          field.name,
+          String(field.value ?? ''),
+          field.type === 'textarea'
+        );
     }
   }
   return entryList;